fix(fw): cast uint32_t args to unsigned long for %lu format
CodeQL cpp/wrong-type-format-argument at MQTTClient.cpp:892 - uptime (uint32_t)
passed to %lu. Same latent mismatch at :686 (_retryInterval) and :858
(WATCHDOG_MS/1000, uptime). Cast all three to match the already-cast sites in
the file. Benign on ESP32 (both 32-bit); silences the checker. No behaviour change.

INVARIANT_OK: 1

+++ lib/thesada-core/src/MQTTClient.cpp
@@ -683,7 +683,7 @@ void MQTTClient::connect() {
     char err[64];
     // TODO: migrate to structured logging
     snprintf(err, sizeof(err), "Failed (rc=%d) - retry in %lums",
-             _client.state(), _retryInterval);
+             _client.state(), (unsigned long)_retryInterval);
     Log::warn(TAG, err);
     _retryCount++;
     _retryInterval = min(_retryInterval * 2, (uint32_t)RETRY_MAX_MS);
@@ -855,7 +855,7 @@ void MQTTClient::loop() {
       uint32_t uptime = (now - _connectedSinceMs) / 1000;
       char wmsg[80];
       // TODO: migrate to structured logging
-      snprintf(wmsg, sizeof(wmsg), "Watchdog: no activity for %lus - forcing reconnect (up %lus)", WATCHDOG_MS / 1000, uptime);
+      snprintf(wmsg, sizeof(wmsg), "Watchdog: no activity for %lus - forcing reconnect (up %lus)", (unsigned long)(WATCHDOG_MS / 1000), (unsigned long)uptime);
       Log::warn(TAG, wmsg);
       _client.disconnect();
       _wifiClient.stop();
@@ -889,7 +889,7 @@ void MQTTClient::loop() {
     uint32_t uptime = (millis() - _connectedSinceMs) / 1000;
     char dmsg[64];
     // TODO: migrate to structured logging
-    snprintf(dmsg, sizeof(dmsg), "Connection lost after %lu seconds", uptime);
+    snprintf(dmsg, sizeof(dmsg), "Connection lost after %lu seconds", (unsigned long)uptime);
     Log::warn(TAG, dmsg);
     _connectedSinceMs = 0;
   }
